typecast tolower/toupper arguments to int to make picky compilers complain less

+++ lib/url.c
@@ -2326,7 +2326,7 @@ static CURLcode CreateConnection(struct SessionHandle *data,
 
 	/* Now, build <protocol>_proxy and check for such a one to use */
 	while(*protop)
-	  *envp++ = tolower(*protop++);
+	  *envp++ = tolower((int)*protop++);
 
 	/* append _proxy */
 	strcpy(envp, "_proxy");
@@ -2349,7 +2349,7 @@ static CURLcode CreateConnection(struct SessionHandle *data,
 	if(!prox && !strequal("http_proxy", proxy_env)) {
           /* There was no lowercase variable, try the uppercase version: */
 	  for(envp = proxy_env; *envp; envp++)
-	    *envp = toupper(*envp);
+	    *envp = toupper((int)*envp);
 	  prox=curl_getenv(proxy_env);
 	}
 
@@ -2578,7 +2578,7 @@ static CURLcode CreateConnection(struct SessionHandle *data,
     if(type) {
       char command;
       *type=0;                     /* it was in the middle of the hostname */
-      command = toupper(type[6]);
+      command = toupper((int)type[6]);
       switch(command) {
       case 'A': /* ASCII mode */
 	data->set.ftp_ascii = 1;
